api/requirements.txt
Flask
Flask-Cors
firebase-admin
google-cloud-storage
google-cloud-vision
pyzbar
Pillow
gunicorn
google-genai
google
python-dotenv